gwspamd: web: Add Chart.js Minified Library to Assets
In this commit, we added the Chart.js minified library to our project's
assets directory for local use. This will allow us to use Chart.js
without relying on a CDN (Content Delivery Network) and provide better
performance for our users.

By adding the minified version, we reduced the size of the library,
which will result in faster page load times. Additionally, having the
library stored locally means that we don't need to make a network
request to load the library, which will further improve the performance
of our application.

Overall, this commit should provide better performance and a better
user experience for our project.

Signed-off-by: Memet Zx <zxce3@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20230428110727.331117-2-zxce3@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>