gwspamd: web: Refactor Profile Update Page
The previous code had a table layout for the form, which has been
replaced with a better-structured HTML form layout. The form now
includes separate input fields for first name, last name, username,
email, and password, and it has a new title and styling. The image
upload button and modal have also been updated to match the new design.

Signed-off-by: Memet Zx <zxce3@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20230428193458.366546-3-zxce3@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>